refactor: replace models/keys/mnemonic with lumos hd
