+++ packages/neuron-wallet/src/database/chain/entities/hd-public-key-info.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, PrimaryGeneratedColumn, Index, CreateDateColumn } from 'typeorm'
 import HdPublicKeyInfoModel from '../../../models/keys/hd-public-key-info'
-import { AddressType } from '../../../models/keys/address'
+import { AddressType } from '@ckb-lumos/hd'
 
 @Entity()
 export default class HdPublicKeyInfo {
