+++ packages/neuron-wallet/src/models/address.ts
@@ -1,4 +1,4 @@
-import { AddressType } from '../models/keys/address'
+import { AddressType } from '@ckb-lumos/hd'
 
 export enum AddressVersion {
   Testnet = 'testnet',
