+++ packages/neuron-wallet/src/services/hardware/ledger.ts
@@ -7,7 +7,8 @@ import type Transport from '@ledgerhq/hw-transport'
 import { Observable, timer } from 'rxjs'
 import { takeUntil, filter, scan } from 'rxjs/operators'
 import Transaction from '../../models/chain/transaction'
-import Address, { AddressType } from '../../models/keys/address'
+import { AddressType } from '@ckb-lumos/hd'
+import Address from '../../models/keys/address'
 import logger from '../../utils/logger'
 import NetworksService from '../../services/networks'
 import { generateRPC } from '../../utils/ckb-rpc'
