+++ packages/neuron-wallet/src/services/wallets.ts
@@ -4,7 +4,8 @@ import Store from '../models/store'
 import Keystore from '../models/keys/keystore'
 import WalletDeletedSubject from '../models/subjects/wallet-deleted-subject'
 import { WalletListSubject, CurrentWalletSubject } from '../models/subjects/wallets'
-import { AccountExtendedPublicKey, DefaultAddressNumber } from '../models/keys/key'
+import { AccountExtendedPublicKey } from '../models/keys/key'
+import { DefaultAddressNumber } from '../models/keys/address'
 import { Address as AddressInterface } from '../models/address'
 
 import FileService from './file'
