+++ packages/neuron-wallet/tests/block-sync-renderer/full-synchronizer.test.ts
@@ -1,6 +1,6 @@
 import { scriptToAddress } from '../../src/utils/scriptAndAddress'
 import { when } from 'jest-when'
-import { AddressType } from '../../src/models/keys/address'
+import { AddressType } from '@ckb-lumos/hd'
 import { Address, AddressVersion } from '../../src/models/address'
 import SystemScriptInfo from '../../src/models/system-script-info'
 import FullSynchronizer from '../../src/block-sync-renderer/sync/full-synchronizer'
