+++ packages/neuron-wallet/tests/services/address.test.ts
@@ -2,7 +2,7 @@ import { AccountExtendedPublicKey } from '../../src/models/keys/key'
 import SystemScriptInfo from '../../src/models/system-script-info'
 import { OutputStatus } from '../../src/models/chain/output'
 import OutputEntity from '../../src/database/chain/entities/output'
-import { AddressType } from '../../src/models/keys/address'
+import { AddressType } from '@ckb-lumos/hd'
 import { Address } from '../../src/models/address'
 import Transaction from '../../src/database/chain/entities/transaction'
 import { TransactionStatus } from '../../src/models/chain/transaction'
