+++ packages/neuron-wallet/tests/services/tx/transaction-sender.test.ts
@@ -188,7 +188,7 @@ import Input from '../../../src/models/chain/input'
 import Script, { ScriptHashType } from '../../../src/models/chain/script'
 import Output from '../../../src/models/chain/output'
 import Keystore from '../../../src/models/keys/keystore'
-import { AddressType } from '../../../src/models/keys/address'
+import { AddressType } from '@ckb-lumos/hd'
 import WitnessArgs from '../../../src/models/chain/witness-args'
 import CellWithStatus from '../../../src/models/chain/cell-with-status'
 import SystemScriptInfo from '../../../src/models/system-script-info'
